add OV

+++ Common/GraphicsBase.cpp
@@ -312,7 +312,7 @@ Text::Text(const Window & window_, const std::string & path_to_ttf_, unsigned in
                 texture,
                 glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
                 glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-                face->glyph->advance.x
+                static_cast<GLuint>(face->glyph->advance.x)
         };
         Characters.insert(std::pair<GLchar, Character>(c, character));
     }
